+++ CHANGELOG.md
@@ -5,9 +5,23 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.0.12] - March 10, 2020
+
+### Added
+
+- `markdown` module with `codeBlock` function
+
+### Changed
+
+- Bumped a number of dep versions.
+- Reuses `exec.ExecOptions` type for `execAndCapture` options so you can pass
+  through all the things.
+
 ## [0.0.11] - Feb 20, 2020
 
 ### Added
+
+---
 
 - `isAnnotationLevel` type guard for GitHub annotation levels.